format_infer: move abstract format
