+++ fpy2/backend/mpfx/__init__.py
@@ -2,7 +2,7 @@
 MPFX backend: compiler to C++ number library.
 """
 
+from ...analysis.format_infer import AbstractFormat, SupportedContext
 from .compiler import MPFXCompiler, MPFXCompileError
 from .elim_round import ElimRound
-from .format import AbstractFormat, SupportedContext
 from .format_infer import FormatInfer, FormatAnalysis
